add fr code for m31
