+++ code_producers/src/wasm_elements/m31/fr-data.wat
@@ -1,0 +1,16 @@
+;;(data (i32.const 0) "\a0\01\00\00")
+(data (i32.const 40) "@\00\00\00\00\00\00\00")
+(data (i32.const 56) "\01\00\00\00\00\00\00\00")
+(data (i32.const 176) "\01\00\00\00\ff\ff\ff\ff")
+(data (i32.const 184) "\ff\ff\ff\ff\00\00\00\00")
+(data (i32.const 192) "\01\00\00\00\fe\ff\ff\ff")
+(data (i32.const 200) "\01\00\00\00\00\00\00\00")
+(data (i32.const 208) "\ff\ff\ff\ff\00\00\00\00")
+(data (i32.const 216) "\00\00\00\00\00\00\00\00")
+(data (i32.const 224) "\00\00\00\80\ff\ff\ff\7f")
+(data (i32.const 232) "\01\00\00\80\ff\ff\ff\7f")
+(data (i32.const 240) "\f9\ff\ff\ff\06\00\00\00")
+(data (i32.const 248) "\ff\ff\ff\ff\00\00\00\00")
+(data (i32.const 256) "\98NQ\0d\8b\87X\da")
+(data (i32.const 264) "\00\00\00\80\00\00\00\00")
+(data (i32.const 408) "\00\00\00\80\ff\ff\ff\7f")
