+++ code_producers/src/wasm_elements/m31/fr-types.wat
@@ -1,0 +1,11 @@
+(type $_sig_i32i32 (func  (param i32 i32)))
+(type $_sig_i32 (func  (param i32)))
+(type $_sig_i32ri32 (func  (param i32) (result i32)))
+(type $_sig_i32i32ri32 (func  (param i32 i32) (result i32)))
+(type $_sig_i32i32i32ri32 (func  (param i32 i32 i32) (result i32)))
+(type $_sig_i32i32i32 (func  (param i32 i32 i32)))
+(type $_sig_i32i64i32 (func  (param i32 i64 i32)))
+(type $_sig_i32i64 (func  (param i32 i64)))
+(type $_sig_i32i32i32i32 (func  (param i32 i32 i32 i32)))
+(type $_sig_i64i64ri64 (func  (param i64 i64) (result i64)))
+(type $_sig_i32i32ri64 (func  (param i32 i32) (result i64)))
