--- examples/multiplier2/multiplier2.sym
@@ -1,3 +1,0 @@
-1,1,0,main.c
-2,2,0,main.a
-3,3,0,main.b
