--- examples/multiplier2/multiplier2_constraints.json
@@ -1,5 +1,0 @@
-{
-"constraints": [
-[{"2":"18446744069414584320"},{"3":"1"},{"1":"18446744069414584320"}]
-]
-}
