--- examples/multiplier2/multiplier2_cpp/Makefile
@@ -1,22 +1,0 @@
-CC=g++
-CFLAGS=-std=c++11 -O3 -I.
-DEPS_HPP = circom.hpp calcwit.hpp fr.hpp
-DEPS_O = main.o calcwit.o fr.o fr_asm.o
-
-ifeq ($(shell uname),Darwin)
-	NASM=nasm -fmacho64 --prefix _
-endif
-ifeq ($(shell uname),Linux)
-	NASM=nasm -felf64
-endif
-	
-all: multiplier2
-	
-%.o: %.cpp $(DEPS_HPP)
-	$(CC) -c $< $(CFLAGS)
-
-fr_asm.o: fr.asm
-	$(NASM) fr.asm -o fr_asm.o
-	
-multiplier2: $(DEPS_O) multiplier2.o
-	$(CC) -o multiplier2 *.o -lgmp
